Implement downloading episodes
The simple solution of using DownloadManager doesn't work because it doesn't follow redirects. Duck taping over that with HEAD requests following the redirects doesn't work either because for some god forsaken reason some Headgum podcasts involve a redirect that only responds 302 for Get and responds 200 for HEAD. The final solution was to use exoplayer's offline functionality.