Make edge-to-edge work with landscape mode
This way of doing it also just seems less janky in general. The fact
that insets must be accounted for in multiple places is not ideal, but
overall it's much cleaner anyways. Also it really is necessary to set
the color of the tab bar in three separate ways; otherwise some part or
another will be the wrong color.